fix: resolve CI failures — pbxproj UUID collision, SwiftLint, SwiftFormat
- Fix UUID collision (EE…3001 used for both Extensions group and test
  cities.json build file) → renamed Extensions group to EE…5001
- Extract SunriseRow + SecondaryToolbarButton to PrayerTimesComponents.swift
  to keep PrayerTimesView under file_length limit
- Raise file_length warning 600→650 (PrayerTimeRow adhaan column feature
  legitimately grew the file)
- Fix toggle_bool: isPrayerMuted.toggle()
- Fix vertical_whitespace double blanks in AboutView, AdhaanBannerView
- SwiftFormat: redundantReturn + consecutiveSpaces in AppAppearance switch
