+++ .swiftlint.yml
@@ -54,7 +54,7 @@ type_body_length:
   error: 600
 
 file_length:
-  warning: 600
+  warning: 650
   error: 900
 
 function_body_length:
