+++ iqamah/Views/PrayerTimesView.swift
@@ -254,7 +254,11 @@ struct PrayerTimesTable: View {
                         ),
                         isHighlighted: isNextPrayer(adjustedTime: adjusted),
                         isPickerExpanded: expandedPrayerName == prayer.name,
-                        onTogglePicker: { withAnimation(.easeInOut(duration: 0.18)) { expandedPrayerName = expandedPrayerName == prayer.name ? nil : prayer.name } },
+                        onTogglePicker: {
+                            withAnimation(.easeInOut(duration: 0.18)) {
+                                expandedPrayerName = expandedPrayerName == prayer.name ? nil : prayer.name
+                            }
+                        },
                         onAdjust: { delta in adjustPrayerTime(for: prayer.name, delta: delta) }
                     )
                 }
@@ -316,39 +320,6 @@ struct PrayerTimesTable: View {
             }
         }
         return adjustedTime == self.adjustedTime(for: (name: "Fajr", time: prayerTimes.fajr))
-    }
-}
-
-// MARK: - Sunrise Row (US-0028)
-
-/// Muted info row for Sunrise — not a prayer, no adjustment controls.
-struct SunriseRow: View {
-    let time: Date
-    let formatter: DateFormatter
-
-    var body: some View {
-        HStack(spacing: 16) {
-            HStack(spacing: 14) {
-                Image(systemName: "sunrise.fill")
-                    .font(.callout)
-                    .foregroundColor(.secondary) // AC-0063: no opacity reduction on semantic colour
-                    .frame(width: 44, height: 36)
-                Text("Sunrise")
-                    .font(.callout)
-                    .foregroundColor(.secondary)
-            }
-            Spacer()
-            Text(formatter.string(from: time))
-                .font(.callout)
-                .foregroundColor(.secondary)
-                .monospacedDigit()
-                .frame(minWidth: 100, alignment: .trailing)
-            Color.clear.frame(width: 76)
-        }
-        .padding(.horizontal, 20)
-        .padding(.vertical, 10)
-        .accessibilityElement(children: .ignore)
-        .accessibilityLabel("Sunrise at \(formatter.string(from: time))")
     }
 }
 
@@ -504,7 +475,7 @@ struct PrayerTimeRow: View {
                     }
 
                     // Per-prayer mute
-                    Button(action: { isPrayerMuted = !isPrayerMuted }) {
+                    Button(action: { isPrayerMuted.toggle() }) {
                         Image(systemName: isPrayerMuted ? "speaker.slash.fill" : "speaker.wave.2.fill")
                             .font(.callout)
                             .foregroundStyle(isPrayerMuted ? .orange : .secondary)
@@ -613,55 +584,21 @@ struct PrayerTimeRow: View {
         .accessibilityElement(children: .contain)
         .accessibilityLabel(
             "\(name) at \(formatter.string(from: time))"
-            + (adjustment != 0 ? ", adjusted \(adjustment) min" : "")
-            + (isPrayerMuted ? ", muted" : "")
-            + (isHighlighted ? ", next prayer" : "")
+                + (adjustment != 0 ? ", adjusted \(adjustment) min" : "")
+                + (isPrayerMuted ? ", muted" : "")
+                + (isHighlighted ? ", next prayer" : "")
         )
     }
 
     private var iconName: String {
         switch name {
-        case "Fajr":    "sun.horizon.fill"
+        case "Fajr": "sun.horizon.fill"
         case "Sunrise": "sunrise.fill"
-        case "Dhuhr":   "sun.max.fill"
-        case "Asr":     "sun.min.fill"
+        case "Dhuhr": "sun.max.fill"
+        case "Asr": "sun.min.fill"
         case "Maghrib": "sunset.fill"
-        case "Isha":    "moon.stars.fill"
-        default:        "clock.fill"
+        case "Isha": "moon.stars.fill"
+        default: "clock.fill"
         }
-    }
-}
-
-// MARK: - Secondary toolbar button
-
-/// Flat toolbar-style button used in the secondary bar below the primary header.
-/// Matches macOS convention: no border, subtle background on hover only.
-private struct SecondaryToolbarButton: View {
-    let label: String
-    let systemImage: String
-    let action: () -> Void
-
-    @State private var isHovering = false
-
-    var body: some View {
-        Button(action: action) {
-            HStack(spacing: 5) {
-                Image(systemName: systemImage)
-                    .font(.system(size: 12, weight: .medium))
-                Text(label)
-                    .font(.system(size: 12, weight: .medium))
-            }
-            .foregroundColor(isHovering ? .primary : .secondary)
-            .padding(.horizontal, 10)
-            .padding(.vertical, 6)
-            .background(
-                RoundedRectangle(cornerRadius: 6, style: .continuous)
-                    .fill(isHovering
-                        ? Color(nsColor: .quaternaryLabelColor).opacity(0.5)
-                        : Color.clear)
-            )
-        }
-        .buttonStyle(.plain)
-        .onHover { isHovering = $0 }
     }
 }
